Extract form validation types from sendForm for better reuse
